:white_check_mark: 가게 목록 조회 테스트 추가 #LF1-274

+++ src/main/java/kr/bb/store/domain/store/handler/response/SimpleStorePagingResponse.java
@@ -13,5 +13,5 @@ import java.util.List;
 @AllArgsConstructor
 public class SimpleStorePagingResponse {
     private List<SimpleStoreResponse> simpleStores;
-    private Integer totalCnt;
+    private Long totalCnt;
 }
